Split some of the test code into Utils
