+++ test/fsatracetest.cabal
@@ -25,5 +25,6 @@ test-suite fsatrace-test
                   , extra
                   , process
                   , transformers
+    other-modules: Utils
     hs-source-dirs: .
     default-language:    Haskell2010
